// mcp-servers/gatling/src/apiClientGenerated/gatlingEnterpriseSchemas.ts
export type RunRequest = {
  /**
   * Non-empty trimmed string
   */
  title?: string;
  /**
   * Non-empty trimmed string
   */
  description?: string;
};